MOF-127 - Finalized show on map feature

+++ src/js/components/user/Profile.svelte
@@ -3,20 +3,35 @@ import { onDestroy, onMount } from "svelte";
 import { fade } from "svelte/transition";
 import InfiniteScroll from "svelte-infinite-scroll";
 import Spinner from "./../elements/Spinner.svelte";
-import { isEmpty } from "./../../utils/commonUtils.js";
 import CustomSelect from "../elements/CustomSelect.svelte";
-import { isLoggedIn, userData, selectedUserProfileData } from "../../store";
+import {
+  isLoggedIn,
+  isShowOnMapMode,
+  userData,
+  markersStore,
+  openedMarkerData,
+  selectedUserProfileData,
+  selectedYear,
+  shouldDisplayShowOnMap,
+} from "../../store";
 import ButtonPrimary from "../elements/ButtonPrimary.svelte";
 import Modal from "../Modal.svelte";
 import EditSpot from "./EditSpot.svelte";
 import DeleteSpot from "./DeleteSpot.svelte";
 import Popup from "../Popup.svelte";
+import { getProfileYears } from "./../../utils/datesUtils.js";
 import {
+  isEmpty,
   loadFromLocalStorage,
   removeFromLocalStorage,
 } from "../../utils/commonUtils";
 import { getUserSpots } from "../../api/spot";
-import { EMPTY_YEAR_STRING, MAX_SPOTS_PER_PAGE } from "../../constants";
+import {
+  ALL_YEARS_STRING,
+  EMPTY_YEAR_STRING,
+  MAX_SPOTS_PER_PAGE,
+} from "../../constants";
+import { permalink } from "../../utils/mapUtils/permalink";
 
 export let onAddSpotBtnClick;
 export let showUserProfile;
@@ -53,21 +68,7 @@ const fetchSpots = ({ year, offset, isNewFetch = false }) => {
       const { spots, years } = result;
       if (isNewFetch) spotsList = [];
       newBatch = spots ? [...spots] : [];
-      yearsToApply = years
-        ? [
-            ...new Set(
-              Object.values(years)
-                .map((y) => (y === null ? EMPTY_YEAR_STRING : y))
-                .filter((y) => y)
-                .sort(
-                  (a, b) =>
-                    (a === EMPTY_YEAR_STRING) - (b === EMPTY_YEAR_STRING) ||
-                    -(a > b) ||
-                    +(a < b)
-                )
-            ),
-          ]
-        : [];
+      yearsToApply = getProfileYears(years);
       if (currentYear === undefined || year === undefined) {
         currentYear = yearsToApply[0];
       }
@@ -84,6 +85,7 @@ const fetchSpots = ({ year, offset, isNewFetch = false }) => {
 
 onMount(() => {
   fetchSpots({});
+  shouldDisplayShowOnMap.set(false);
 });
 
 onDestroy(() => {
@@ -107,6 +109,11 @@ const handleAddSpot = () => {
 const handleYearSelect = (event) => {
   const { value } = event.detail.detail;
   currentYear = value !== EMPTY_YEAR_STRING ? value : "";
+  offset = 0;
+  if (currentYear === ALL_YEARS_STRING) {
+    fetchSpots({ isNewFetch: true });
+    return;
+  }
   fetchSpots({ year: `${currentYear}`, isNewFetch: true });
 };
 
@@ -122,11 +129,60 @@ const handleDelete = (spot) => {
 
 const onLoadMore = () => {
   offset += MAX_SPOTS_PER_PAGE;
+  if (currentYear === ALL_YEARS_STRING) {
+    fetchSpots({ offset });
+    return;
+  }
   fetchSpots({ year: `${currentYear}`, offset });
 };
 
 const onSubmitChanges = () => {
   fetchSpots({ year: `${currentYear}`, isNewFetch: true });
+};
+
+const onSpotClick = (spot) => {
+  if (isCurrentUser) {
+    return;
+  }
+  const {
+    id,
+    artistCrew,
+    spotStatus: status,
+    description,
+    img,
+    title,
+    videoLink: video,
+    publicBanner: { banner, bannerUrl },
+    location: { lat, lng },
+    year,
+    link,
+  } = spot;
+  openedMarkerData.set({
+    id,
+    artistCrew,
+    status,
+    description,
+    img: { src: img, title: title || id },
+    video,
+    user: $selectedUserProfileData,
+    firm: { banner, bannerUrl },
+    coords: { lat, lng },
+    year,
+    link,
+  });
+  if (currentYear !== ALL_YEARS_STRING) {
+    shouldDisplayShowOnMap.set(true);
+    markersStore.set({ spots: spotsList, years: yearsToApply });
+  }
+  permalink.update({ params: { marker: id } });
+  showUserProfile(false);
+};
+
+const handleShowOnMapClick = () => {
+  markersStore.set({ spots: spotsList, years: yearsToApply });
+  isShowOnMapMode.set(true);
+  selectedYear.set(`${currentYear}`);
+  showUserProfile(false);
 };
 </script>
 
@@ -138,8 +194,9 @@ const onSubmitChanges = () => {
     {#if isCurrentUser}
       <button type="button" class="logout" on:click={handleLogout}
         >Logout</button>
-    {:else}
-      <button type="button" class="show-on-map">Show on map</button>
+    {:else if !isLoading && currentYear !== ALL_YEARS_STRING}
+      <button type="button" class="show-on-map" on:click={handleShowOnMapClick}
+        >Show on map</button>
     {/if}
   </div>
   {#if !!spotsList.length || isShowSpinner}
@@ -156,26 +213,28 @@ const onSubmitChanges = () => {
       {#if !isLoading}
         <div class="spots">
           {#each spotsList as spot}
-            <div class="spot-card">
+            <div class="spot-card" on:click={() => onSpotClick(spot)}>
               <img
                 loading="lazy"
                 src={spot.thumbnail}
                 alt={spot.title || `${username}'s art`}
                 in:fade />
-              <div class="overlay">
-                <button
-                  type="button"
-                  class="edit"
-                  on:click={() => handleEdit(spot)} />
-                <button
-                  type="button"
-                  class="delete"
-                  on:click={() => handleDelete(spot)} />
-              </div>
+              {#if isCurrentUser}
+                <div class="overlay">
+                  <button
+                    type="button"
+                    class="edit"
+                    on:click={() => handleEdit(spot)} />
+                  <button
+                    type="button"
+                    class="delete"
+                    on:click={() => handleDelete(spot)} />
+                </div>
+              {/if}
             </div>
           {/each}
           <InfiniteScroll
-            hasMore={newBatch.length}
+            hasMore={newBatch.length === MAX_SPOTS_PER_PAGE}
             threshold={100}
             on:loadMore={onLoadMore}
             elementScroll={document.querySelector(".modal")} />
@@ -217,18 +276,20 @@ const onSubmitChanges = () => {
 <style lang="scss">
 .container {
   display: flex;
+  flex: 1 0 auto;
   flex-direction: column;
   align-items: center;
-  flex: 1 0 auto;
   width: 100%;
   max-width: 938px;
 }
+
 .top {
   display: flex;
   align-self: stretch;
   justify-content: space-between;
   margin-bottom: 36px;
 }
+
 .username {
   color: var(--color-dark);
   font-size: 24px;
@@ -236,10 +297,11 @@ const onSubmitChanges = () => {
   line-height: 1.22;
   text-transform: uppercase;
 }
+
 .logout {
+  transition: opacity 0.3s;
   border: 0;
   opacity: 0.4;
-  transition: opacity 0.3s;
   background: none;
   color: var(--color-accent);
   font-size: 18px;
@@ -247,17 +309,31 @@ const onSubmitChanges = () => {
   line-height: 1.22;
   text-transform: uppercase;
   cursor: pointer;
+
   &:hover {
     opacity: 1;
   }
 }
 
+.show-on-map {
+  padding: 0;
+  border: 0;
+  background: none;
+  color: var(--color-accent);
+  font-size: 18px;
+  font-weight: 600;
+  line-height: 22px;
+  cursor: pointer;
+}
+
 .empty-state {
   margin: auto 0;
   text-align: center;
+
   img {
     width: 195px;
   }
+
   p {
     margin: 36px 0;
     color: var(--color-dark);
@@ -281,38 +357,43 @@ const onSubmitChanges = () => {
 
 .spots {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(275px, 1fr));
   grid-auto-rows: 160px;
   grid-gap: 4vmin;
+  grid-template-columns: repeat(auto-fill, minmax(275px, 1fr));
   justify-content: space-between;
   width: 100%;
 }
 
 .spot-card {
   position: relative;
-  border-radius: 2px;
   overflow: hidden;
+  border-radius: 2px;
   background-color: #d3d3d3;
+  cursor: pointer;
+
   img {
     width: 100%;
     height: 100%;
     object-fit: cover;
+
     user-select: none;
   }
+
   .overlay {
     display: flex;
-    align-items: center;
-    justify-content: center;
     position: absolute;
     top: 0;
     right: 0;
     bottom: 0;
     left: 0;
+    align-items: center;
+    justify-content: center;
+    transition: opacity 0.3s;
     border-radius: inherit;
     opacity: 0;
-    transition: opacity 0.3s;
     background: rgba($color: #000, $alpha: 0.45);
   }
+
   &:hover {
     .overlay {
       opacity: 1;
@@ -320,16 +401,15 @@ const onSubmitChanges = () => {
   }
 }
 
-.edit,
-.delete {
-  border: 0;
-  background-color: var(--color-accent);
-  background-position: 50% 50%;
-  background-repeat: no-repeat;
+.edit, .delete {
   width: 54px;
   height: 54px;
   margin: 12px;
+  border: 0;
   border-radius: 2px;
+  background-color: var(--color-accent);
+  background-repeat: no-repeat;
+  background-position: 50% 50%;
   cursor: pointer;
 }
 
@@ -349,14 +429,24 @@ const onSubmitChanges = () => {
 
 @media (max-width: 767px) {
   .top {
+    position: relative;
     flex-direction: column-reverse;
+    margin-bottom: 18px;
   }
+
   .logout {
     width: 30px;
     height: 30px;
     margin: -30px 0 52px;
-    font-size: 0;
     background: url(../../../images/logout.svg) 50% 50%/27px 27px no-repeat;
+    font-size: 0;
+  }
+
+  .show-on-map {
+    position: absolute;
+    top: 54px;
+    right: 0;
   }
 }
+
 </style>
